+++ app/_ui/chat-wrapper.tsx
@@ -3,14 +3,14 @@
 import ChatForm from "./chat-form";
 import Conversation from "./conversation";
 import React, { useState, useEffect, Fragment } from "react";
-import { Chat, ChatResponse, ChatUserSession } from "../lib/model";
+import { Chat, ChatResponse, ChatUserSession } from "../_lib/model";
 import Sidebar from "./sidebar";
 import Detail from "./detail";
 import UtilityIconRow from "./utility-icon-row";
 import { Transition } from "@headlessui/react";
 import { SettingsHeading } from "./typography";
-import { fetchHistory } from "../lib/api";
-import { History } from "../lib/model";
+import { fetchHistory } from "../_lib/api";
+import { History } from "../_lib/model";
 import { MaterialSymbol } from "react-material-symbols";
 import "react-material-symbols/outlined";
