+++ app/_ui/detail.tsx
