+++ app/settings/history/page.tsx
@@ -1,6 +1,6 @@
-import HistoryList from "@/app/ui/history-list";
-import { SettingsHeading } from "@/app/ui/typography";
-import PersonaSkeleton from "@/app/ui/settings/persona-skeleton";
+import HistoryList from "@/app/_ui/history-list";
+import { SettingsHeading } from "@/app/_ui/typography";
+import PersonaSkeleton from "@/app/_ui/settings/persona-skeleton";
 import { Suspense } from "react";
 
 export default function HistoryPage() {
